Bail early when not rendering in the browser
Prior to this change, the use of this plugin was assumed to only happen
in the browser. However, when generating static pages from the
tiddlywiki command line these macros are still executed.

Considering the Disqus comments are useless in a static rendered page
anyway we can simply not render the comments or counts in these cases.

This change bails out of the macro early if it detects that is is not in
a browser.

The original error message prior to this fix was:

    $:/plugins/bimlas/disqus/macro/counter:28
                    if (document.getElementById("dsq-count-scr") === null) {
                    ^

    ReferenceError: document is not defined
        at MacroCallWidget.exports.run ($:/plugins/bimlas/disqus/macro/counter:28:3)
        at MacroCallWidget.Widget.evaluateMacroModule ($:/core/modules/widgets/widget.js:199:21)
        at MacroCallWidget.Widget.getVariableInfo ($:/core/modules/widgets/widget.js:117:14)
        at MacroCallWidget.execute ($:/core/modules/widgets/macrocall.js:52:23)
        at MacroCallWidget.render ($:/core/modules/widgets/macrocall.js:32:7)
        at $:/core/modules/widgets/widget.js:373:15
        at Object.$tw.utils.each (/Users/suki/source/dev-tritarget-org/node_modules/tiddlywiki/boot/boot.js:69:12)
        at ElementWidget.Widget.renderChildren ($:/core/modules/widgets/widget.js:372:12)
        at ElementWidget.render ($:/core/modules/widgets/element.js:49:7)
        at $:/core/modules/widgets/widget.js:373:15

https://github.com/Jermolene/TiddlyWiki5/blob/master/boot/bootprefix.js#L24