Add disqus-did-insert-element event
Prior to this change, there was no way for other JS/modules to know when
a Disqus thread was rendered. I discovered this need when I mad a button
that would reveal the comment thread but it needed to scroll to the
Disqus element since the tiddler was several scroll views away. I
realized that an easy way to accomplish this was to have the macro
announce when it inserted the Disqus DIV.

This change will dispatch a disqus-did-insert-element event that other
JS can tap into. Since the actual DIV is returned as a string to be
rendered I had to defer the event. And since TiddlyWiki rendering is
synchronous using the nextTick utility would allow for the DOM to be
there before the event is dispatched.

For an example of how this might be used:

    $tw.rootWidget.addEventListener(
      'disqus-did-insert-element',
      function(event) {
        $tw.pageScroller.scrollIntoView(event.target);
      }
    );

To see this in action visit https://tritarget.org/ with code examples at
https://github.com/sukima/dev-tritarget-org/commit/d9d722b95480c69a1bbf5d6196dc8df1d4dde238